requirements.txt
openai
python-dotenv
ollama
faiss-cpu
chromadb
tiktoken
langchain
langchain_community
langchain_chroma
langchain_nvidia_ai_endpoints
